Create w2e8_solana_programs.md
Solana Programs
Using the examples in the repo

Modify the existing msg! in example1-helloworld to log the program ID.
Retrieve the total program size of example1-helloworld.
Retrieve the lamport balance of example2-counter.
Modify the client for example2-counter to feed an incorrect address for the greeting account to the program.